Control.Arrow.Transformer.Static

Description

Arrow transformer adding static information.

Documentation

newtype StaticArrow f a b c

An arrow type that augments the underlying arrow with static information.

Constructors

StaticArrow (f (a b c))

type StaticMonadArrow m = StaticArrow (WrappedMonad m)

A special case is monads applied to the whole arrow, in contrast to Kleisli arrows, in which the monad is applied to the output.

type StaticArrowArrow a s = StaticArrow (WrappedArrow a s)

A special case.

wrap :: (Applicative f, Arrow a) => f (a b c) -> StaticArrow f a b c

unwrap :: (Applicative f, Arrow a) => StaticArrow f a b c -> f (a b c)

wrapA :: (Arrow a, Arrow a') => a s (a' b c) -> StaticArrowArrow a s a' b c

unwrapA :: (Arrow a, Arrow a') => StaticArrowArrow a s a' b c -> a s (a' b c)

wrapM :: (Monad m, Arrow a) => m (a b c) -> StaticMonadArrow m a b c

unwrapM :: (Monad m, Arrow a) => StaticMonadArrow m a b c -> m (a b c)