Control.Arrow.Transformer.Stream

Description

Arrow transformer lifting an arrow to streams.

Documentation

newtype StreamArrow a b c

Arrows between streams.

Note: lift is only a functor if *** in the underlying arrow is.

Constructors

StreamArrow (a (Stream b) (Stream c))

runStream :: ArrowLoop a => StreamArrow a (e, b) c -> a (e, Stream b) (Stream c)

Run a stream processor on a stream of inputs, obtaining a stream of outputs.

Typical usage in arrow notation:

   proc p -> do
       ...
       ys <- (|runStream (\x -> ...)|) xs

Here xs refers to the input stream and x to individual elements of that stream. ys is bound to the output stream.

type StreamMap = StreamArrow (->)

Mappings of streams

type StreamMapST s = StreamArrow (Kleisli (ST s))

In-place state updates.

Note: this is an arrow type, and lift can be used to promote arrows from Kleisli (ST s): the resulting arrow updates the state for each stream element in turn, and as long as the final state in not required all is well. However, lift does not preserve composition, because this monad isn't commutative. In particular, a composition of lifts of state transformers will not work, as the second will require the final state of the first.

runStreamST :: (forall s. StreamMapST s e c) -> StreamMap e c

Encapsulate a local state.

class (ArrowCircuit a, Arrow a') => ArrowAddStream a a' | a -> a' where

Adding a StreamArrow to an arrow type, but not necessarily as the outer arrow transformer.

Typically a composite arrow type is built by applying a series of arrow transformer to a base arrow (usually either a function arrow or a Kleisli arrow. One can add a transformer to the top of this stack using the lift method of the ArrowTransformer class, or remove a state transformer from the top of the stack using the runStream encapsulation operator. The methods of this class add and remove state transformers anywhere in the stack. In the instance

   instance Arrow a => ArrowAddStream (ArrowStream a) a

they are equivalent to lift and runStream respectively. Instances are lifted through other transformers with

   instance ArrowAddStream a a' =>
       ArrowAddStream (FooArrow a) (FooArrow a')

Methods

liftStream :: a' e b -> a e b

Lift a computation from an arrow to a stream processing one.

Typical usage in arrow notation:

   proc p -> ...
       (|liftStream cmd|)

elimStream :: a (e, b) c -> a' (e, Stream b) (Stream c)

Run a stream processor on a stream of inputs, obtaining a stream of outputs.

Typical usage in arrow notation:

   proc p -> do
       ...
       ys <- (|elimStream (\x -> ...)|) xs

Here xs refers to the input stream and x to individual elements of that stream. ys is bound to the output stream.