Control.Arrow.Transformer.Error

Description

An arrow transformer that adds error handling.

TODO: the operations here are inconsistent with other arrow transformers.

Documentation

newtype ErrorArrow ex a b c

An arrow that augments an existing arrow with possible errors. The ArrowError class contains methods for raising and handling these errors.

Constructors

ErrorArrow (a b (Either ex c))

runError

Arguments

:: ArrowChoice a

=> ErrorArrow ex a e b
computation that may raise errors

-> a (e, ex) b
computation to handle errors

-> a e b

Encapsulate an error-raising computation, by completely handling any errors.

Typical usage in arrow notation:

proc p -> ...
		body `runError` \ex -> handler

class (ArrowError ex a, Arrow a') => ArrowAddError ex a a' | a -> a' where

Adding a ErrorArrow to an arrow type, but not necessarily as the outer arrow transformer.

Typically a composite arrow type is built by applying a series of arrow transformer to a base arrow (usually either a function arrow or a Kleisli arrow. One can add a transformer to the top of this stack using the lift method of the ArrowTransformer class, or remove a state transformer from the top of the stack using the runError encapsulation operator. The methods of this class add and remove state transformers anywhere in the stack. In the instance

instance Arrow a => ArrowAddError ex (ArrowError ex a) a

they are equivalent to lift and runError respectively. Instances are lifted through other transformers with

instance ArrowAddError ex a a' =>
		ArrowAddError ex (FooArrow a) (FooArrow a')

This could be combined with handle, since the resulting arrow is always the arrow of the handler. Separating them has the advantage of consistency with the other arrows, and might give more helpful type error messages.

Minimal complete definition

liftError, elimError

Methods

liftError :: a' e b -> a e b

Lift a computation from an arrow to one with error handling.

Typical usage in arrow notation:

proc p -> ...
		(|liftError cmd|)

elimError :: a e b -> a' (e, ex) b -> a' e b

Elimination of errors from a computation, by completely handling any errors.

Typical usage in arrow notation:

proc p -> ...
		body `elimError` \ex -> handler