Control.Arrow.Transformer.Reader

Description

Arrow transformer that adds a read-only state (i.e. an environment).

Documentation

newtype ReaderArrow r a b c

An arrow type that augments an existing arrow with a read-only state (or environment). The ArrowReader class contains the operations on this state.

Constructors

ReaderArrow (a (b, r) c)

runReader :: Arrow a => ReaderArrow r a e b -> a (e, r) b

Encapsulation of a state-reading computation, taking a value for the state.

Typical usage in arrow notation:

proc p -> ...
		(|runReader cmd|) env

class (ArrowReader r a, Arrow a') => ArrowAddReader r a a' | a -> a' where

Adding a ReaderArrow to an arrow type, but not necessarily as the outer arrow transformer.

Typically a composite arrow type is built by applying a series of arrow transformer to a base arrow (usually either a function arrow or a Kleisli arrow. One can add a transformer to the top of this stack using the lift method of the ArrowTransformer class, or remove a state transformer from the top of the stack using the runReader encapsulation operator. The methods of this class add and remove state transformers anywhere in the stack. In the instance

instance Arrow a => ArrowAddReader r (ArrowReader r a) a

they are equivalent to lift and runReader respectively. Instances are lifted through other transformers with

instance ArrowAddReader r a a' =>
		ArrowAddReader r (FooArrow a) (FooArrow a')

Minimal complete definition

liftReader, elimReader

Methods

liftReader :: a' e b -> a e b

Lift a computation from an arrow to one with an added environment.

Typical usage in arrow notation:

proc p -> ...
		(|liftReader cmd|)

elimReader :: a e b -> a' (e, r) b

Elimination of a state reader from a computation, taking a value for the state.

Typical usage in arrow notation:

proc p -> ...
		(|elimReader cmd|) env